consolidate get-invalid-visits and purge-invalid-visits
