+++ server/seeders/purge-invalid-visits.ts
@@ -1,5 +1,9 @@
 /* eslint-disable no-await-in-loop, no-console */
 
+import fs from 'fs';
+
+import validate from 'validator';
+
 import { QueryInterface, QueryTypes } from 'sequelize';
 
 interface Visit {
@@ -12,41 +16,59 @@ interface Visit {
 
 export default {
   up: async (queryInterface: QueryInterface): Promise<void> => {
-    const transaction = await queryInterface.sequelize.transaction();
+    const visitsCountQuery = await queryInterface.sequelize.query(
+      'SELECT COUNT(*) FROM visit',
+      {
+        type: QueryTypes.SELECT,
+      }
+    );
+    const { count: visitsCount } = visitsCountQuery[0] as { count: number };
+
+    const MAX_ROWS = 10000;
+    const invalidVisits: Visit[] = [];
+    let offset = 0;
 
     try {
-      const badVisitsQuery = await queryInterface.sequelize.query(
-        `SELECT id FROM visit
-        WHERE url LIKE '%localhost%' OR
-        (url LIKE '%blogspot.com%' AND url LIKE '%preview%') OR
-        url LIKE 'http:///' OR
-        url LIKE 'http://appyet_base/%' OR
-        id = 55664 OR
-        id = 1319147 OR
-        id = 976248 OR
-        id = 4357709 OR
-        id = 1862496`,
-        {
-          type: QueryTypes.SELECT,
-          transaction,
-        }
-      );
-      const badVisits = badVisitsQuery as Visit[];
-      const badVisitIds = badVisits.map(({ id }) => id);
+      while (offset < visitsCount) {
+        const visitsQuery = await queryInterface.sequelize.query(
+          `SELECT * FROM visit ORDER BY id LIMIT ${MAX_ROWS} OFFSET ${offset}`,
+          {
+            type: QueryTypes.SELECT,
+          }
+        );
+        const visits = visitsQuery as Visit[];
 
-      await queryInterface.sequelize.query(
-        'DELETE FROM visit WHERE id IN (?)',
-        {
-          type: QueryTypes.DELETE,
-          replacements: [badVisitIds],
-          transaction,
-        }
-      );
+        visits.forEach((visit) => {
+          const validCanonical = validate.isURL(visit.canonical);
+          const validUrl =
+            validate.isURL(visit.url) &&
+            !validate.contains(visit.url, 'localhost');
 
-      transaction.commit();
-      console.log(`Success! Deleted ${badVisits.length} rows`);
+          if (!validCanonical || !validUrl) {
+            invalidVisits.push(visit);
+          }
+        });
+
+        offset += MAX_ROWS;
+      }
+
+      if (invalidVisits.length) {
+        console.log(`There are ${invalidVisits.length} invalid visits`);
+        fs.writeFileSync('invalid-visits.json', JSON.stringify(invalidVisits));
+
+        const invalidVisitIds = invalidVisits.map(({ id }) => id);
+
+        await queryInterface.sequelize.query(
+          'DELETE FROM visit WHERE id IN (?)',
+          {
+            type: QueryTypes.DELETE,
+            replacements: [invalidVisitIds],
+          }
+        );
+      }
+
+      console.log('Success!');
     } catch (error) {
-      transaction.rollback();
       console.error(error);
     }
   },
